Find a better "local" psysh install if possible.
Use a globally installed psysh bin script or phar as a "launcher" for a local version, if possible. Starting from the current working directory, check for 'psy/psysh' as a composer project or dependency. If found, load that one instead of the globally installed version.

Note that this only works with local versions new enough to include the extracted \Psy\bin helper (i.e. local versions newer than this commit). It does provide useful feedback for how to fix the issue or work around it, and in practice shouldn't affect too many people too much.