Work around unclosed single quoted string.
For some reason, the parse error isn't "unexpected EOF" with an unclosed single quote string, unlike just about every other unclosed statement. So we'll add a special case, just for this.

Unfortunately, it's not perfect. For example, the statement `$foo 'bar` can't be valid, so it _should_ throw a parse exception instead of continuing on the next line. I can't find a way to target just that case without messing up other unclosed single quote strings.

It works great in a bunch of other cases, though, so this is probably okay. Maybe.

Fixes #181