Add a `useUnicode` config option.
This is a workaround for things like #270, because it's easier to stick to ASCII than to fight with Windows code page issues.

Eventually, the manual config won't really be necessary, because we'll detect when we're going to run into issues and disable unicode output ourselves. But that's someday, and this fix is today.